fragment 코드로 리펙토링
