+++ presentation/src/main/java/com/team_ia/idea_archive_android/ui/main/AuthenticationSuccessActivity.kt
@@ -1,4 +1,16 @@
 package com.team_ia.idea_archive_android.ui.main
 
-class AuthenticationSuccessActivity {
+import com.team_ia.idea_archive_android.R
+import com.team_ia.idea_archive_android.databinding.ActivityAuthenticationSuccessPageBinding
+import com.team_ia.idea_archive_android.ui.base.BaseActivity
+
+class AuthenticationSuccessActivity : BaseActivity<ActivityAuthenticationSuccessPageBinding>(R.layout.activity_authentication_success_page) {
+    override fun createView() {
+
+    }
+
+    override fun observeEvent() {
+
+    }
+
 }
