+++ presentation/src/main/res/layout/fragment_auth_code_input_page.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="authCodeInput"
-            type="com.team_ia.idea_archive_android.ui.main.AuthCodeInputActivity" />
+            type="com.team_ia.idea_archive_android.ui.main.AuthCodeInputFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -286,7 +286,6 @@
             android:text="@string/check"
             android:textColor="@color/white"
             android:textSize="18sp"
-            android:onClick="@{authCodeInput::onClick}"
             app:layout_constraintBottom_toBottomOf="@id/g_check_btn_b"
             app:layout_constraintEnd_toEndOf="@id/g_check_btn_e"
             app:layout_constraintStart_toStartOf="@id/g_check_btn_s"
